+++ CHANGELOG.md
@@ -15,7 +15,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Allow Jenkins job notifications to reach lists. Before, such mails were rejected due to being "auto-submitted".
 * Do not recognize sudo messages as automated message (#248)
 * Fixed using x-attach-listkey with emails from Thunderbird that include protected headers.
-
+* Ensure ASCII-8BIT as external encoding, this should ensure that plain text emails in different charsets can be parsed (#409)
 
 ## [3.4.1] / 2019-09-16